backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
spacy
rdflib
rapidfuzz
langdetect
argostranslate
torch
transformers
sentence-transformers
faiss-cpu
numpy
groq
tiktoken
datasets
evaluate
